[ProductAttribute] Fixes after PR reviews
